update routes and migration

+++ database/migrations/2026_05_29_081011_fix_user_id_on_observations_table.php
@@ -6,20 +6,19 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     */
     public function up(): void
-{
-    Schema::table('observations', function (Blueprint $table) {
-        $table->dropForeign(['user_id']);
-        $table->dropColumn('user_id');
-    });
-}
+    {
+        Schema::table('observations', function (Blueprint $table) {
+            if (Schema::hasColumn('observations', 'user_id')) {
+                try {
+                    $table->dropForeign(['user_id']);
+                } catch (\Exception $e) {
+                }
+                $table->dropColumn('user_id');
+            }
+        });
+    }
 
-    /**
-     * Reverse the migrations.
-     */
     public function down(): void
     {
         //
